add the shopping cart and change the floats to decimals.

+++ src/Ecommerce.Shop/Products/Domain/Product.cs
@@ -7,10 +7,10 @@ namespace Ecommerce.Shop.Products.Domain
         public string Name { get; private set; }
         public string Description { get; set; }
         public Category Category { get; set; }
-        public float Price { get; private set; }
-        public float DiscountPrice { get; private set; }
+        public decimal Price { get; private set; }
+        public decimal DiscountPrice { get; private set; }
 
-        public Product(string name, float price)
+        public Product(string name, decimal price)
         {
             Name = name;
             Price = price;
